Add visual test cases page to process flow
