+++ packages/showcase/components/TestCases/ProcessFlow/case01.js
@@ -32,18 +32,17 @@ const data1 = {
   ]
 }
 
-const data2: any = {
+const data2 = {
   journeys: [
-    { path: ["1", "2", "3", "4"], size: 1500 },
-    { path: ["1", "2", "3", "5", "4"], size: 1200 },
-    { path: ["1", "7", "5", "8"], size: 700 },
+    { path: ["1", "10", "2", "3", "4"], size: 1500 },
+    { path: ["1", "10", "2", "3", "5", "4"], size: 1200 },
+    { path: ["9", "7", "5", "8"], size: 700 },
     { path: ["9", "2", "3", "8"], size: 600 },
-    { path: ["1", "2", "5", "6", "11"], size: 230 },
+    { path: ["1", "10", "2", "5", "6", "11"], size: 230 },
     { path: ["1", "2", "3", "6", "11"], size: 130 },
     { path: ["1", "2", "3", "4", "11"], size: 290 },
-    { path: ["1", "2", "3", "12", "11"], size: 120 },
-    { path: ["1", "2", "3", "4", "13"], size: 620 },
-    { path: ["4"], size: 23 }
+    { path: ["1", "10", "2", "3", "12", "11"], size: 120 },
+    { path: ["1", "10", "2", "3", "4", "13"], size: 620 }
   ],
   nodes: [
     { id: "1", group: "start" },
@@ -55,47 +54,63 @@ const data2: any = {
     { id: "7" },
     { id: "8", group: "end" },
     { id: "9", group: "start" },
+    { id: "10" },
     { id: "11", group: "end" },
     { id: "12" },
     { id: "13", group: "end" }
   ]
 }
 
-const marathon = ({ test, afterAll, container }: IMarathon): void => {
+const data3 = {
+  journeys: [
+    { path: ["1", "10", "2", "3", "4"], size: 1500 },
+    { path: ["1", "10", "2", "3", "5", "4"], size: 1200 },
+    { path: ["9", "7", "5", "8"], size: 700 },
+    { path: ["9", "2", "3", "8"], size: 600 }
+  ],
+  nodes: [
+    { id: "1", group: "start" },
+    { id: "2" },
+    { id: "3" },
+    { id: "4" },
+    { id: "5" },
+    { id: "7" },
+    { id: "8", group: "end" },
+    { id: "9", group: "start" },
+    { id: "10" }
+  ]
+}
+
+export const marathon = ({ test, afterAll, container }) => {
   const viz = new ProcessFlow(container)
 
-  test("Focuses a link", () => {
+  test("Renders a process flow with no data", () => {
+    viz.draw()
+  })
+
+  test("Renders a process flow with an empty dataset", () => {
+    viz.data({})
+    viz.draw()
+  })
+
+  test("Adds data", () => {
     viz.data(data1)
-    viz.config({
-      focusElement: {
-        type: "link",
-        matchers: { sourceId: "3", targetId: "5" }
-      }
-    })
-    viz.accessors("node", {
-      label: (node: any) => `Node ${node.id}`
-    })
+    viz.config({ duration: 2e3 })
     viz.draw()
   })
 
-  test("Removes focus", () => {
-    viz.config({ focusElement: {} })
-    viz.draw()
-  })
-
-  test("Focuses a node", () => {
+  test("Updates the data", () => {
     viz.data(data2)
-    viz.config({
-      focusElement: {
-        type: "node",
-        matchers: { id: "3" }
-      }
-    })
-    viz.accessors("node", {
-      label: (node: any) => `Node ${node.id}`,
-      content: [{ key: "Description", value: "This is a node." }, { key: "Comment", value: "This comment is boring." }]
-    })
     viz.draw()
+  })
+
+  test("Updates the data", () => {
+    viz.data(data3)
+    viz.draw()
+  })
+
+  test("Closes the viz", () => {
+    viz.close()
   })
 
   afterAll(() => {
@@ -103,11 +118,4 @@ const marathon = ({ test, afterAll, container }: IMarathon): void => {
   })
 }
 
-export default props => (
-  <Layout pathname={props.url.pathname}>
-    <Card>
-      <CardHeader>Process Flow Visualization Tests - Element Focussing</CardHeader>
-      <Marathon test={marathon} timeout={2e3} />
-    </Card>
-  </Layout>
-)
+export const title = "Data Updates"
