+++ packages/showcase/components/TestCases/ProcessFlow/case02.js
@@ -32,17 +32,18 @@ const data1 = {
   ]
 }
 
-const data2 = {
+const data2: any = {
   journeys: [
-    { path: ["1", "10", "2", "3", "4"], size: 1500 },
-    { path: ["1", "10", "2", "3", "5", "4"], size: 1200 },
-    { path: ["9", "7", "5", "8"], size: 700 },
+    { path: ["1", "2", "3", "4"], size: 1500 },
+    { path: ["1", "2", "3", "5", "4"], size: 1200 },
+    { path: ["1", "7", "5", "8"], size: 700 },
     { path: ["9", "2", "3", "8"], size: 600 },
-    { path: ["1", "10", "2", "5", "6", "11"], size: 230 },
+    { path: ["1", "2", "5", "6", "11"], size: 230 },
     { path: ["1", "2", "3", "6", "11"], size: 130 },
     { path: ["1", "2", "3", "4", "11"], size: 290 },
-    { path: ["1", "10", "2", "3", "12", "11"], size: 120 },
-    { path: ["1", "10", "2", "3", "4", "13"], size: 620 }
+    { path: ["1", "2", "3", "12", "11"], size: 120 },
+    { path: ["1", "2", "3", "4", "13"], size: 620 },
+    { path: ["4"], size: 23 }
   ],
   nodes: [
     { id: "1", group: "start" },
@@ -54,63 +55,47 @@ const data2 = {
     { id: "7" },
     { id: "8", group: "end" },
     { id: "9", group: "start" },
-    { id: "10" },
     { id: "11", group: "end" },
     { id: "12" },
     { id: "13", group: "end" }
   ]
 }
 
-const data3 = {
-  journeys: [
-    { path: ["1", "10", "2", "3", "4"], size: 1500 },
-    { path: ["1", "10", "2", "3", "5", "4"], size: 1200 },
-    { path: ["9", "7", "5", "8"], size: 700 },
-    { path: ["9", "2", "3", "8"], size: 600 }
-  ],
-  nodes: [
-    { id: "1", group: "start" },
-    { id: "2" },
-    { id: "3" },
-    { id: "4" },
-    { id: "5" },
-    { id: "7" },
-    { id: "8", group: "end" },
-    { id: "9", group: "start" },
-    { id: "10" }
-  ]
-}
-
-const marathon = ({ test, afterAll, container }) => {
+export const marathon = ({ test, afterAll, container }: IMarathon): void => {
   const viz = new ProcessFlow(container)
 
-  test("Renders a process flow with no data", () => {
-    viz.draw()
-  })
-
-  test("Renders a process flow with an empty dataset", () => {
-    viz.data({})
-    viz.draw()
-  })
-
-  test("Adds data", () => {
+  test("Focuses a link", () => {
     viz.data(data1)
-    viz.config({ duration: 2e3 })
+    viz.config({
+      focusElement: {
+        type: "link",
+        matchers: { sourceId: "3", targetId: "5" }
+      }
+    })
+    viz.accessors("node", {
+      label: (node: any) => `Node ${node.id}`
+    })
     viz.draw()
   })
 
-  test("Updates the data", () => {
+  test("Removes focus", () => {
+    viz.config({ focusElement: {} })
+    viz.draw()
+  })
+
+  test("Focuses a node", () => {
     viz.data(data2)
+    viz.config({
+      focusElement: {
+        type: "node",
+        matchers: { id: "3" }
+      }
+    })
+    viz.accessors("node", {
+      label: (node: any) => `Node ${node.id}`,
+      content: [{ key: "Description", value: "This is a node." }, { key: "Comment", value: "This comment is boring." }]
+    })
     viz.draw()
-  })
-
-  test("Updates the data", () => {
-    viz.data(data3)
-    viz.draw()
-  })
-
-  test("Closes the viz", () => {
-    viz.close()
   })
 
   afterAll(() => {
@@ -118,11 +103,4 @@ const marathon = ({ test, afterAll, container }) => {
   })
 }
 
-export default props => (
-  <Layout pathname={props.url.pathname}>
-    <Card>
-      <CardHeader>Process Flow Visualization Tests - Data Updates</CardHeader>
-      <Marathon test={marathon} timeout={2e3} />
-    </Card>
-  </Layout>
-)
+export const title = "Element Focussing"
