+++ packages/showcase/components/TestCases/ProcessFlow/index.js
@@ -1,0 +1,8 @@
+import * as case01 from "./case01"
+import * as case02 from "./case02"
+import * as case03 from "./case03"
+import * as case04 from "./case04"
+import * as case05 from "./case05"
+import * as case06 from "./case06"
+
+export default [case01, case02, case03, case04, case05, case06]
